requirements.txt
streamlit
pandas
sqlalchemy
passlib
python-slugify
python-dotenv